Extracted PreviewColorization logic
